+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
   * [Armory]: More settings and visual improvements
   * [GameMenu]: Some more infos on the GameMenu
+  * [GameMenu]: Fix Option -fix1